Add files via upload
This Python scripts contains the derivation of Linear Regression from scratch. We generally use gradient descent to minimise the loss function. The `LinearRegression` class which I have implemented uses both the Gradient Descent and Normal Equation to find the argmin of the cost function. Since gradient descent is an iterative process, it is computationally expensive when we get hold of a fairly large amount of data. We can use the OLS solver in such a cases. OLS solver uses data and solves it in a much more efficient way.